dagster/requirements.txt
adal==1.2.7
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
anyio==4.4.0
applicationinsights==0.11.10
argcomplete==3.3.0
azure-appconfiguration==1.1.1
azure-batch==14.2.0
azure-cli==2.62.0
azure-cli-core==2.62.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.53
azure-graphrbac==0.60.0
azure-identity==1.17.1
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.0.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==31.0.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==30.0.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.5.1
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datafactory==8.0.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-netapp==10.1.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.4.0
azure-mgmt-redis==14.3.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.0
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.0.0b6
azure-mgmt-servicelinker==1.2.0b2
azure-mgmt-signalr==2.0.0b1
azure-mgmt-sql==4.0.0b17
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==21.2.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.2.0
azure-monitor-query==1.2.0
azure-multiapi-storage==1.2.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.19.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
backoff==2.2.1
bcrypt==4.1.3
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
croniter==2.0.5
cryptography==42.0.8
dagster==1.7.12
dagster-cloud==1.7.12
dagster-cloud-cli==1.7.12
dagster-databricks==0.23.12
dagster-graphql==1.7.12
dagster-pipes==1.7.12
dagster-pyspark==0.23.12
dagster-spark==0.23.12
dagster-webserver==1.7.12
databricks-api==0.9.0
databricks-cli==0.18.0
databricks-sdk==0.17.0
# Editable install with no version control (ddb==0.0.0)
-e /home/kader/dagster/dagster
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
fabric==3.2.2
filelock==3.15.4
fsspec==2024.6.1
github3.py==4.0.1
google-auth==2.31.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.62.2
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
install==1.3.5
invoke==2.2.0
isodate==0.6.1
javaproperties==0.5.2
Jinja2==3.1.4
jmespath==1.0.1
jsondiff==2.0.0
knack==0.11.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.28.1
msal-extensions==1.2.0b1
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.0.5
oauthlib==3.2.2
packaging==24.1
paramiko==3.4.0
pendulum==3.0.0
pex==2.9.0
pip-system-certs==4.0
pkginfo==1.11.1
pluggy==1.5.0
portalocker==2.10.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==5.9.8
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycomposefile==0.0.31
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
PySocks==1.7.1
pyspark==3.5.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
scp==0.13.6
semver==2.13.0
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sshtunnel==0.1.5
starlette==0.37.2
structlog==24.2.0
tabulate==0.9.0
time-machine==2.14.2
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.3.3
websockets==12.0
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
